backend_sam2/requirements.txt
fastapi
uvicorn
python-multipart
ultralytics
numpy
pillow
opencv-python-headless
